% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{CleanText}
\alias{CleanText}
\title{CleanText}
\usage{
CleanText(
  TrainData = NULL,
  ValidationData = NULL,
  TestData = NULL,
  TextColumn = NULL,
  MergeColumns = NULL,
  RemovePunctuation = TRUE,
  StopWords = "none",
  StopWordsSource = "stopwords-iso"
)
}
\arguments{
\item{TrainData}{Training data. If ValidationData or TestData is supplied, w2v will be created on TrainData only and applied to ValidationData and TestData.}

\item{ValidationData}{Validation data from a modeling context. TextColumnNames must match TrainData.}

\item{TestData}{Test data from a modeling context. TextColumnNames must match TrainData.}

\item{TextColumn}{Character vector of column names to utilize}

\item{MergeColumns}{Columns used to join to TrainData (and the others). Data is split during stopword removal and must be joined back to source data.}

\item{RemovePunctuation}{Logical or vector of logical that is the same length as TextColumnNames}

\item{StopWords}{Default "none". For stopwords package usage supply language code e.g. "en" for english https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes. Otherwise, supply a character vector of stopwords. If you want to supply a modifed version of what's returned from the stopwords package, run this and modify as you wish, stopwords::stopwords(language = 'en', source = 'stopwords-iso')}

\item{StopWordsSource}{Default "stopwords-iso" and only applies if you supply a language code in the StopWords parameter. Other options include "snowball", "stopwords-iso", "misc", "smart", "marimo", "ancient", "nltk", "perseus".}
}
\description{
Remove punctuation and stopwords
}
\author{
Adrian Antico
}
\concept{Word2Vec}
