% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{LexicalDiversity}
\alias{LexicalDiversity}
\title{LexicalDiversity}
\usage{
LexicalDiversity(
  dt,
  TextColumns = NULL,
  Measures = "TTR",
  RemoveSymbols = TRUE,
  RemoveHyphens = TRUE,
  RemovePunctuation = TRUE,
  RemoveNumbers = TRUE,
  LogBase = 10,
  MATTR_Window = 100L,
  MSTTR_Segment = 100L
)
}
\arguments{
\item{dt}{data.table}

\item{TextColumns}{Names of text columns to analyze}

\item{Measures}{Default is "TTR". Also available "C", "R", "CTTR", "U", "S", "K", "I", "D", "Vm", "Maas", "MATTR", "MSTTR"}

\item{RemoveSymbols}{Logical}

\item{RemoveHyphens}{Logical}

\item{RemovePunctuation}{Logical}

\item{RemoveNumbers}{Logical}

\item{LogBase}{Default 10}

\item{MATTR_Window}{Numeric value defining the size of the moving average window for computation of the Moving-Average Type-Token Ration}

\item{MSTTR_Segment}{Numeric value defining the size of each segment for the computation of the Mean Segmental Type-Token Ratio}
}
\description{
Generate lexical divrsity stats
}
\seealso{
Other NLP Stats: 
\code{\link{Readability}()},
\code{\link{Sentiment}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
