% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{N_Grams}
\alias{N_Grams}
\title{N_Gram}
\usage{
N_Grams(
  dt,
  dt_type = "raw",
  TextColumns = NULL,
  IDcols = NULL,
  N = 2,
  StopWords = "en",
  StopWordsSource = "stopwords-iso"
)
}
\arguments{
\item{dt}{data.table}

\item{dt_type}{Default "raw" which means non-tokenized text columns. "tokenized" means the dt text column has already been tokenized}

\item{N}{The N in N-Gram}

\item{StopWords}{Default "none". For stopwords package usage supply language code e.g. "en" for english https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes. Otherwise, supply a character vector of stopwords. If you want to supply a modifed version of what's returned from the stopwords package, run this and modify as you wish, stopwords::stopwords(language = 'en', source = 'stopwords-iso')}

\item{StopWordsSource}{Default "stopwords-iso" and only applies if you supply a language code in the StopWords parameter. Other options include "snowball", "stopwords-iso", "misc", "smart", "marimo", "ancient", "nltk", "perseus".}
}
\description{
Generate n-gram columns from text columns
}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{Readability}()},
\code{\link{Sentiment}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
