% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{Readability}
\alias{Readability}
\title{Readability}
\usage{
Readability(
  dt,
  TextColumns = NULL,
  Measures = "Flesch",
  RemoveHyphens = TRUE,
  MinSentenceLength = 1,
  MaxSentenceLength = 10000,
  Intermediate = TRUE
)
}
\arguments{
\item{dt}{data.table}

\item{TextColumns}{Names of text columns to analyze}

\item{Measures}{Default is "Flesch". Also available "ARI", "Bormuth.MC", "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau.ECP", "Coleman.Liau.grade", "Coleman.Liau.short", "Dale.Chall", "Danielson.Bryan", "Dickes.Steiwer", "DRP", "ELF", "Farr.Jenkins.Paterson", "Flesch.PSK", "Flesch.Kincaid", "FOG", "FOG.PSK", "FOG.NRI", "FORCAST", "Fucks", "Linsear.Write", "LIW", "nWS", "nWS.2", "nWS.3", "nWS4", "RIX", "Scrabble", "SMOG", "SMOG.C", "SMOG.simple", "SMOG.de", "Spache", "Spache.old", "Strain", "Traenkle.Bailer", "Wheeler.Smith", "meanSentenceLength", "meanWordSyllables"}

\item{RemoveHyphens}{TRUE. FALSE to not remove them}

\item{MinSentenceLength}{Defautl 1}

\item{MaxSentenceLength}{Default 10000}

\item{Intermediate}{Logical. TRUE to include intermediate quantities in the output}
}
\description{
Generate readability stats
}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{N_Grams}()},
\code{\link{Sentiment}()},
\code{\link{TextColsDistance}()},
\code{\link{TextColsSimilarity}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
