% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{Sentiment}
\alias{Sentiment}
\title{Sentiment}
\usage{
Sentiment(
  dt,
  TextColumns = NULL,
  Response = "numeric",
  CombineTextGroupVar = NULL,
  Language = "english",
  RemoveStopWords = TRUE,
  Stemming = TRUE
)
}
\arguments{
\item{dt}{data.table}

\item{TextColumns}{Names of text columns to analyze}

\item{Response}{Default "numeric" is a sentiment score. "Binary" returns either "positive" or "negative". "Direction" is Binary along with a "neutral" response.}

\item{CombineTextGroupVar}{A factor variable by which documents can be grouped. This helpful when joining e.g. news from the same day or move reviews by the same author}

\item{Language}{Default "english"}

\item{RemoveStopWords}{Logical}

\item{Stemming}{Logical}

\item{Measures}{Default is "Flesch". Also available "ARI", "Bormuth.MC", "Bormuth.GP", "Coleman", "Coleman.C2", "Coleman.Liau.ECP", "Coleman.Liau.grade", "Coleman.Liau.short", "Dale.Chall", "Danielson.Bryan", "Dickes.Steiwer", "DRP", "ELF", "Farr.Jenkins.Paterson", "Flesch.PSK", "Flesch.Kincaid", 'FOG", "FOG.PSK", "FOG.NRI", "FORCAST", "Fucks", "Linsear.Write", "LIW", "nWS", "nWS.2", "nWS.3", "nWS4", "RIX", "Scrabble", "SMOG", "SMOG.C", "SMOG.simple", "SMOG.de", "Spache", "Spache.old", "Strain", "Traenkle.Bailer", "Wheeler.Smith", "meanSentenceLength", "meanWordSyllables"}
}
\description{
Generate readability stats
}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{N_Grams}()},
\code{\link{Readability}()},
\code{\link{TextColsDistance}()},
\code{\link{TextColsSimilarity}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
