% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{TextColsSimilarity}
\alias{TextColsSimilarity}
\title{TextColsSimilarity}
\usage{
TextColsSimilarity(dt, TextCol1 = NULL, TextCol2 = NULL)
}
\arguments{
\item{dt}{data.table}

\item{TextCol1}{Text column 1}

\item{TextCol2}{Text column 2}
}
\description{
Generate similarity metrics between two text columns
}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{N_Grams}()},
\code{\link{Readability}()},
\code{\link{Sentiment}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
