% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{TextSummary}
\alias{TextSummary}
\title{TextSummary}
\usage{
TextSummary(dt = NULL, TextColumns = NULL, RemoveStats = NULL)
}
\arguments{
\item{dt}{data.table}

\item{RemoveStats}{NULL. If you want any metrics suppressed, supply as a character vector. Metrics include, "document", "chars", "sents", "tokens", "types", "puncts", "numbers", "symbols", "urls", "tags", "emojis"}
}
\description{
Generate text summary stats
}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{Readability}()},
\code{\link{Sentiment}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
