% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{TextColsDistance}
\alias{TextColsDistance}
\title{TextColsDistance}
\usage{
TextColsDistance(
  dt,
  TextCol1 = NULL,
  TextCol2 = NULL,
  Margin = "documents",
  DistanceMeasure = "euclidean",
  MinkowskiPower = 2
)
}
\arguments{
\item{dt}{data.table}

\item{TextCol1}{Text column 1}

\item{TextCol2}{Text column 2}

\item{Margin}{"documents" or "features"}

\item{DistanceMeasure}{"euclidean", "manhattan", "maximum", "canberra", "minkowski"}

\item{MinkowskiPower}{Default 2. Only used when DistanceMeasure is "minkowski"}
}
\description{
Generate distance metrics between two text columns
}
\examples{
\dontrun{
dt <- AutoQuant::FakeDataGenerator(N=1000, AddComment = TRUE)
dt1 <- AutoQuant::FakeDataGenerator(N=1000, AddComment = TRUE)
dt[, Comment2 := dt1$Comment]
}

}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{N_Grams}()},
\code{\link{Readability}()},
\code{\link{Sentiment}()},
\code{\link{TextColsSimilarity}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
