% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP_Analysis.R
\name{TextColsSimilarity}
\alias{TextColsSimilarity}
\title{TextColsSimilarity}
\usage{
TextColsSimilarity(
  dt,
  TextCol1 = NULL,
  TextCol2 = NULL,
  Margin = "documents",
  Method = "cosine",
  MaxSampleSize = 10000
)
}
\arguments{
\item{dt}{data.table}

\item{TextCol1}{Text column 1}

\item{TextCol2}{Text column 2}

\item{Margin}{"documents" or "features"}

\item{Method}{"correlation", "cosine", "jaccard", "ejaccard", "dice", "edice", "hamann", "simple matching"}
}
\description{
Generate similarity metrics between two text columns
}
\examples{
\dontrun{
dt <- AutoNLP::FakeDataGenerator(N=1000, AddComment = TRUE)
dt1 <- AutoNLP::FakeDataGenerator(N=1000, AddComment = TRUE)
dt[, Comment2 := dt1$Comment]
TextCol1 <- "Comment"
TextCol2 <- "Comment2"
}

}
\seealso{
Other NLP Stats: 
\code{\link{LexicalDiversity}()},
\code{\link{Readability}()},
\code{\link{Sentiment}()},
\code{\link{TextColsDistance}()},
\code{\link{TextSummary}()}
}
\author{
Adrian Antico
}
\concept{NLP Stats}
